fix(proxy): move rust control-plane endpoints into dedicated package
