+++ litellm/proxy/rust_control_plane_endpoints/__init__.py
@@ -1,0 +1,1 @@
+"""Endpoints consumed by the Rust data-plane gateway."""
